:truck: Adaptor 안에 Web이 있어야 했는데, 왜 adaptor.infra 안에 있었나?
