--- src/main/java/com/promptoven/settlementservice/adaptor/infrastructure/web/controller/vo/in/UpdateAccountRequestVO.java
@@ -1,16 +1,0 @@
-package com.promptoven.settlementservice.adaptor.infrastructure.web.controller.vo.in;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UpdateAccountRequestVO {
-
-    private String profileID;
-    private String accountID;
-    private String bankName;
-
-}
