--- src/main/java/com/promptoven/settlementservice/adaptor/infrastructure/web/controller/vo/in/UpdateAddressRequestVO.java
@@ -1,16 +1,0 @@
-package com.promptoven.settlementservice.adaptor.infrastructure.web.controller.vo.in;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UpdateAddressRequestVO {
-
-    private String profileID;
-    private String postcode;
-    private String address;
-
-}
