--- src/main/java/com/promptoven/settlementservice/adaptor/infrastructure/web/controller/vo/in/UpdatePhoneRequestVO.java
@@ -1,15 +1,0 @@
-package com.promptoven.settlementservice.adaptor.infrastructure.web.controller.vo.in;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UpdatePhoneRequestVO {
-
-    private String profileID;
-    private String phone;
-
-}
