--- src/main/java/com/promptoven/settlementservice/adaptor/infrastructure/web/controller/vo/in/UpdateTaxIDRequestVO.java
@@ -1,15 +1,0 @@
-package com.promptoven.settlementservice.adaptor.infrastructure.web.controller.vo.in;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UpdateTaxIDRequestVO {
-
-    private String profileID;
-    private String taxID;
-
-}
