+++ src/main/java/com/promptoven/settlementservice/adaptor/web/controller/vo/in/UpdatePhoneRequestVO.java
@@ -1,0 +1,15 @@
+package com.promptoven.settlementservice.adaptor.web.controller.vo.in;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UpdatePhoneRequestVO {
+
+	private String profileID;
+	private String phone;
+
+}
