+++ src/main/java/com/promptoven/settlementservice/adaptor/web/controller/vo/in/UpdateTaxIDRequestVO.java
@@ -1,0 +1,15 @@
+package com.promptoven.settlementservice.adaptor.web.controller.vo.in;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UpdateTaxIDRequestVO {
+
+	private String profileID;
+	private String taxID;
+
+}
